pyproject.toml
[project]
name = "ag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "autogen-agentchat>=0.7.5",
    "autogen-ext[azure,openai]>=0.7.5",
    "chromadb>=1.4.0",
    "dashscope>=1.25.2",
    "gradio>=6.0.1",
    "moviepy>=2.2.1",
    "openai>=1.0.0",
    "pyjianyingdraft>=0.2.5",
    "python-dotenv>=1.0.1",
    "scenedetect>=0.6.7.1",
    "zai-sdk>=0.2.0",
]

